+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * Fixed a permissions issue with `/sf charge`
 * Fixed #3053
 * Fixed #3075
+* Fixed recipe types showing missing string message
 
 ## Release Candidate 23 (19 May 2021)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#23